+++ company.el
@@ -1054,16 +1054,33 @@ means that `company-mode' is always turned on except in `message-mode' buffers."
                       (const :tag "Except" not)
                       (repeat :inline t (symbol :tag "mode")))))
 
+(defcustom company-global-minibuffer t
+  "Non-nil to enable `company-mode' in the minibuffer.
+The value can be t (meaning only enable if the minibuffer has a local
+`completion-at-point-functions' value) or a custom predicate function."
+  :type 'boolean)
+
 ;;;###autoload
 (define-globalized-minor-mode global-company-mode company-mode company-mode-on)
 
 (defun company-mode-on ()
+  (when (and company-global-minibuffer
+             (minibufferp))
+    (add-hook 'minibuffer-setup-hook #'company--minibuffer-on 100))
   (when (and (not (or noninteractive (eq (aref (buffer-name) 0) ?\s)))
              (cond ((eq company-global-modes t)
                     t)
                    ((eq (car-safe company-global-modes) 'not)
                     (not (memq major-mode (cdr company-global-modes))))
                    (t (memq major-mode company-global-modes))))
+    (company-mode 1)))
+
+(defun company--minibuffer-on ()
+  (when (and (or resize-mini-windows
+                 (not (try-completion "company-pseudo-tooltip" company-frontends)))
+             (if (eq company-global-minibuffer t)
+                 (local-variable-p 'completion-at-point-functions)
+               (funcall company-global-minibuffer)))
     (company-mode 1)))
 
 (defsubst company-assert-enabled ()
@@ -4326,11 +4343,19 @@ Returns a negative number if the tooltip should be displayed above point."
         (overlay-put ov 'company-height height))))
 
 (defun company-pseudo-tooltip-show-at-point (pos column-offset)
+  (when (and (= (window-size) 1)
+             (minibufferp))
+    (add-hook 'window-size-change-functions #'company-pseudo-tooltip-refresh nil t))
   (let* ((col-row (company--col-row pos))
          (col (- (car col-row) column-offset)))
     (when (< col 0) (setq col 0))
     (company--with-face-remappings
      (company-pseudo-tooltip-show (1+ (cdr col-row)) col company-selection))))
+
+(defun company-pseudo-tooltip-refresh (_window)
+  (when company-pseudo-tooltip-overlay
+    (overlay-put company-pseudo-tooltip-overlay 'company-guard nil))
+  (company-pseudo-tooltip-frontend 'post-command))
 
 (defun company-pseudo-tooltip-edit (selection)
   (let* ((height (overlay-get company-pseudo-tooltip-overlay 'company-height))